Move Classes into Separate H and CPP Files
No code added or changed, just moved around.
