+++ LevelInfo.h
@@ -1,0 +1,11 @@
+#pragma once
+
+#include "Usings.h"
+
+struct LevelInfo
+{
+    Price price_;
+    Quantity quantity_;
+};
+
+using LevelInfos = std::vector<LevelInfo>;
