+++ OrderType.h
@@ -1,0 +1,9 @@
+#pragma once
+
+enum class OrderType
+{
+	GoodTillCancel,
+	FillAndKill,
+	FillOrKill,
+	GoodForDay,
+};
