+++ Orderbook.h
@@ -1,0 +1,76 @@
+#pragma once
+
+#include <map>
+#include <unordered_map>
+#include <thread>
+#include <condition_variable>
+#include <mutex>
+
+#include "Usings.h"
+#include "Order.h"
+#include "OrderModify.h"
+#include "OrderbookLevelInfos.h"
+#include "Trade.h"
+
+class Orderbook
+{
+private:
+
+    struct OrderEntry
+    {
+        OrderPointer order_{ nullptr };
+        OrderPointers::iterator location_;
+    };
+
+    struct LevelData
+    {
+        Quantity quantity_{ };
+        Quantity count_{ };
+
+        enum class Action
+        {
+            Add,
+            Remove,
+            Match,
+        };
+    };
+
+    std::unordered_map<Price, LevelData> data_;
+    std::map<Price, OrderPointers, std::greater<Price>> bids_;
+    std::map<Price, OrderPointers, std::less<Price>> asks_;
+    std::unordered_map<OrderId, OrderEntry> orders_;
+    mutable std::mutex ordersMutex_;
+    std::thread ordersPruneThread_;
+    std::condition_variable shutdownConditionVariable_;
+
+    void PruneGoodForDayOrders();
+
+    void CancelOrders(OrderIds orderIds);
+    void CancelOrderInternal(OrderId orderId);
+
+    void OnOrderCancelled(OrderPointer order);
+    void OnOrderAdded(OrderPointer order);
+    void OnOrderMatched(Price price, Quantity quantity, bool isFullyFilled);
+    void UpdateLevelData(Price price, Quantity quantity, LevelData::Action action);
+
+    bool CanFullyFill(Side side, Price price, Quantity quantity) const;
+    bool CanMatch(Side side, Price price) const;
+    Trades MatchOrders();
+
+public:
+
+    Orderbook();
+    Orderbook(const Orderbook&) = delete;
+    void operator=(const Orderbook&) = delete;
+    Orderbook(Orderbook&&) = delete;
+    void operator=(Orderbook&&) = delete;
+    ~Orderbook();
+
+    Trades AddOrder(OrderPointer order);
+    void CancelOrder(OrderId orderId);
+    Trades ModifyOrder(OrderModify order);
+
+    std::size_t Size() const;
+    OrderbookLevelInfos GetOrderInfos() const;
+};
+
