+++ Side.h
@@ -1,0 +1,8 @@
+#pragma once
+
+enum class Side
+{
+    Buy,
+    Sell
+};
+
