+++ TradeInfo.h
@@ -1,0 +1,11 @@
+#pragma once
+
+#include "Usings.h"
+
+struct TradeInfo
+{
+    OrderId orderId_;
+    Price price_;
+    Quantity quantity_;
+};
+
