+++ main.cpp
@@ -1,0 +1,8 @@
+#include "Orderbook.h"
+
+int main()
+{
+    Orderbook orderbook;
+    // Do work.
+    return 0;
+}
